puffin-client: work-around (probable) bug in rkyv
This changes the representation of SimpleMetadata to a sorted
sequence of metadatums. In turns out that we were only using
a BTreeMap to build a de-duplicated sorted sequence anyway.
So we use one during construction, but otherwise convert it to
a Vec before we're done.

We use a new named type, SimpleMetadatum, instead of a tuple
because rkyv is not compatible with tuples when 'strict' mode
is enabled.

The specific bug this works around seems to be related to rkyv's
zero-copy support for BTreeMap. With the BTreeMap, I was hitting
unaligned pointer errors. I stumbled upon this fix by instinct;
I guessed that a zero-copy BTreeMap was a complete and involved
affair, and might be a likely source of the bug. (Because my own
code is of course perfection.)