puffin-client: close the (first) loop on zero-copy deserialization
This commit brings everything from the previous commits together and
actually utilizes zero-copy deserialization for `SimpleMetadata`
(and also `CachePolicy`). MessagePack (via Serde) is still used for
everything else.

The main change is to make our `CachedClient` generic over the
(de)serialization mechanism. That is, we want to continue supporting
Serde while also supporting rkyv without duplicating a bunch of code. To
that end, we introduce a `Cacheable` trait which basically just
describes (de)serialization at a high level: it abstracts over
converting to and from bytes.

The old `CachedClient::get_cached_with_callback` is now just
`get_serde`, and a new method, `get_cacheable`, is introduced that is
more generic. The latter is now used in the registry client for fetching
`SimpleMetadata`.

In ad hoc benchmarks, this seems to be good for about a 10% improvement,
although I've seen it as small as 5%. (The measurements are pretty
noisy.) In particular, this work doesn't fully realize the benefits of
zero-copy deserialization because we are still eagerly converting an
`Archived<SimpleMetadata>` (that's the zero-copy type) to a
`SimpleMetadata` in order to build a `VersionMap`. Indeed, profiles seem
to point to this as one of the bigger bottlenecks in `puffin pip
compile`.

The `DataWithCachePolicy` type has also been rejiggered a bit and beefed
up. Namely, `CachePolicy` is now always zero-copy deserialized (even if
Serde is used for the actual cached response data) and the type
parameter has been removed. That is, where as we previously deserialized
a `CachedPolicy` and the response data in one step, we now do it in two
steps. This is somewhat required by zero-copy deserialization, although
some clever refactoring could probably remove this two-step process if
necessary. The key though here is that other than loading the cache
policy bytes into memory, its deserialization is truly free. Namely, in
the fast path, we make HTTP caching decisions using
`Archived<CachePolicy>`, so in that case, we will never need to
materialize a `CachePolicy` at all. For that reason, I think the
two-step process here is not such a big deal.