puffin-client: actually use SimpleMetadataRaw
This commit introduces SimpleMetadataRaw into the registry
client without actually changing the cache contents. This
effectively is just lining up the API we want. To achieve
this, we actually introduce a perf regression by continuing
to deserialize SimpleMetadata as a MessagePack blob, and then
doing a rkyv serialization followed by a deserialization.

The idea is that we want to change the cached client to deal
with archived types instead of round-tripping through Serde.