pypy-types: introduce lazier 'join_relative' operation
This provides a lazier version of "join relative URL to base URL."
Namely, it avoids needing to parse/create a `BaseUrl` in the case where
the possibly relative URL is actually not relative.

This was requested perhaps for perf reasons, but parsing base URLs
doesn't show up on a profile, so I'm not really sure it's worth doing
for that reason alone. However, this operation is done in a few places,
and having a dedicated function to do it makes the code a bit clearer.

Ref: https://github.com/astral-sh/puffin/pull/1135#discussion_r1468883135