puffin-client: implement HTTP caching semantics
This commit introduces a new module that implements the subset of HTTP
cache semantics that we care about. The implementation was principally
guided by the following things:

1. RFCs 9110 and 9111.
2. The implementation in the `http-cache-semantics` crate.
3. Our zero-copy needs.

The comments in the code should explain things in more detail.

We will actually use this in a subsequent commit.

Note that this brings with it its own Cache-Control header parser. We
have one already (will be deleted in next commit) that was copied out
of the `http-cache-semantics` crate, but it was overall pretty cheesy.
It could also get things wrong with more exotic configurations (how
common those are is unclear).