uv-client: tweak 'fresh_request' to be more flexible
Basically, instead of it wrapping its data up inside an enum, we just
return the data itself. That way, the caller can control what is done
with it. Right now, all uses just wrap it into the enum anyway, so this
looks somewhat pointless, but we'll use this flexibility in the next
commit.